Add jobset-wait command
This command permits to wait for a jobset completion. It first waits
for a new evaluation, and then wait for all job completions. Note if a
job fails, it immediately returns an error.

The main usecase is to get the jobset status in pull requests: CI
runners (gitlab runner or travis) could run this command to get the
jobset status.